requirements.txt
asgiref==3.8.1
Django==5.1.2
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
pillow==10.4.0
psycopg==3.2.3
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
